requirements.txt
PyYAML
requests
beautifulsoup4